+++ cucumber-core/src/main/java/io/cucumber/core/runtime/CucumberExecutionContext.java
@@ -25,6 +25,8 @@ import java.util.function.Consumer;
 
 import static io.cucumber.cienvironment.DetectCiEnvironment.detectCiEnvironment;
 import static io.cucumber.core.exception.ExceptionUtils.printStackTrace;
+import static io.cucumber.core.exception.ExceptionUtils.throwAsUncheckedException;
+import static io.cucumber.core.exception.UnrecoverableExceptions.rethrowIfUnrecoverable;
 import static io.cucumber.messages.Convertor.toMessage;
 import static java.util.Collections.singletonList;
 
@@ -44,6 +46,37 @@ public final class CucumberExecutionContext {
         this.bus = bus;
         this.exitStatus = exitStatus;
         this.runnerSupplier = runnerSupplier;
+    }
+
+    @FunctionalInterface
+    public interface ThrowingRunnable {
+        void run() throws Throwable;
+    }
+
+    public void runInContext(ThrowingRunnable inContextExecution) {
+        startTestRun();
+        execute(() -> {
+            runBeforeAllHooks();
+            inContextExecution.run();
+        });
+        try {
+            execute(this::runAfterAllHooks);
+        } finally {
+            finishTestRun();
+        }
+        Throwable throwable = getThrowable();
+        if (throwable != null) {
+            throwAsUncheckedException(throwable);
+        }
+    }
+
+    private void execute(ThrowingRunnable runnable) {
+        try {
+            runnable.run();
+        } catch (Throwable t) {
+            // Collected in CucumberExecutionContext
+            rethrowIfUnrecoverable(t);
+        }
     }
 
     public void startTestRun() {
